Extracted SequencerImpl out of ConflictCheckingLedgerBridge
