+++ ledger/sandbox-on-x/src/test/suite/scala/com/daml/ledger/sandbox/bridge/SequenceTest.scala
@@ -1,0 +1,13 @@
+package com.daml.ledger.sandbox.bridge
+
+import org.mockito.{ArgumentMatchersSugar, MockitoSugar}
+import org.scalatest.flatspec.AsyncFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+class SequenceTest
+  extends AsyncFlatSpec
+    with MockitoSugar
+    with Matchers
+    with ArgumentMatchersSugar {
+
+}
